修复模型保存
